refactor(measures): move subgraph-gate shared code under _subgraph_gate/
Relocate baseline-store, registry, subgraph-measure contract, and
load-all from packages/measures/src/_shared/measures/ into
packages/measures/src/_subgraph_gate/_internal/. Nothing outside the
subgraph gate consumes these — they're the gate's internal contracts,
not cross-cutting helpers.

Keeping them under _subgraph_gate/ turns ~50 measure→runtime imports
from cross-subdirectory edges (which the maxSubdirCrossRatio axis of
pressure-axes counts) into intra-subdirectory ones, dropping the
measures package's ratio from 0.82 back below the green-CI baseline.
The previously-failing pressure-axes RSCD (1.055) now passes (0.745).
