+++ packages/measures/src/_subgraph_gate/measures/behavioral/ast-purity-ratio.ts
@@ -44,14 +44,14 @@
  * the upper bound once the codebase has been swept once.
  */
 import {Node, SyntaxKind, type SourceFile as MorphSourceFile} from 'ts-morph'
-import {loadBaseline} from '../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'ast-purity-ratio'
