+++ packages/measures/src/_subgraph_gate/measures/behavioral/implicit-globals.ts
@@ -49,14 +49,14 @@
  * (e.g. 12 → 13) gets `fail`).
  */
 import {Node, SyntaxKind, type SourceFile as MorphSourceFile} from 'ts-morph'
-import {loadBaseline} from '../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'implicit-globals'
