+++ packages/measures/src/_subgraph_gate/measures/behavioral/module-state-bindings.ts
@@ -29,14 +29,14 @@
  */
 import {Node, type SourceFile as MorphSourceFile} from 'ts-morph'
 import {communityForFile} from '../../../_shared/community/community-at-depth.ts'
-import {loadBaseline} from '../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'module-state-bindings'
