+++ packages/measures/src/_subgraph_gate/measures/shape/cyclomatic-per-fn.ts
@@ -32,14 +32,14 @@
 // returns nodes from that TS instance, and SyntaxKind constants differ).
 import {ts} from 'ts-morph'
 import {communityForFile} from '../../../_shared/community/community-at-depth.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     Severity,
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 import {isFunctionLikeBoundary, walkFunctions, type FunctionDetail} from './_ast-helpers.ts'
 
 export const CYCLOMATIC_WARN_SOFT = 10
