+++ packages/measures/src/_subgraph_gate/measures/structural/_deferred/modularity-q.ts
@@ -34,14 +34,14 @@
 import {communityAtDepth, siblingGroupParent} from '../../../../_shared/community/community-at-depth.ts'
 import {computeModularityQ} from '../../../../_shared/complexity/hierarchical-complexity-measures.ts'
 import type {Edge, SourceFile} from '../../../../_shared/graph/import-graph.ts'
-import {loadBaseline} from '../../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../../_shared/measures/subgraph-measure.ts'
+} from '../../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'modularity-q'
