+++ packages/measures/src/_subgraph_gate/measures/structural/boundary-width.ts
@@ -34,14 +34,14 @@
 import {readFile} from 'node:fs/promises'
 import {exportedSymbolNames} from '../../../_shared/complexity/exported-symbols.ts'
 import type {SourceFile} from '../../../_shared/graph/import-graph.ts'
-import {loadBaseline} from '../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'boundary-width'
