+++ packages/measures/src/_subgraph_gate/measures/structural/structural-orange.ts
@@ -33,14 +33,14 @@ import {
     computePriorityScoresAtDepth,
     type CommunityPriority,
 } from '../../../_shared/complexity/orange-priority.ts'
-import {loadBaseline} from '../../../_shared/measures/baseline-store.ts'
-import {registerMeasure} from '../../../_shared/measures/registry.ts'
+import {loadBaseline} from '../../_internal/baseline-store.ts'
+import {registerMeasure} from '../../_internal/registry.ts'
 import type {
     SubgraphMeasure,
     SubgraphMeasureInput,
     SubgraphMeasureResult,
     Violation,
-} from '../../../_shared/measures/subgraph-measure.ts'
+} from '../../_internal/subgraph-measure.ts'
 
 export const MEASURE_ID = 'structural-orange'
